# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Server for SSU Schedule Application",
        "title": "SSU Schedule API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1.0",
    "paths": {
        "/departments": {
            "get": {
                "description": "Retrieves SSU departments' list",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "departments"
                ],
                "summary": "get a list of departments",
                "operationId": "get-departments-list",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Departments"
                            }
                        }
                    }
                }
            }
        },
        "/{education_form}/{department}/groups": {
            "get": {
                "description": "Retrieves groups' list based on department and education form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "get a list of groups of a certain department",
                "operationId": "get-groups-list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Education form, e.g. `do`",
                        "name": "education_form",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Department URL, e.g. `knt` for CSIT department",
                        "name": "department",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Groups"
                            }
                        }
                    }
                }
            }
        },
        "/{education_form}/{department}/{group_number}": {
            "get": {
                "description": "Retrieves the schedule based on department, education form and group number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule"
                ],
                "summary": "get the schedule of students for a particular group",
                "operationId": "get-students-schedule",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Education form, e.g. `do`",
                        "name": "education_form",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Department URL, e.g. `knt` for CSIT department",
                        "name": "department",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group number, e.g. `351`",
                        "name": "group_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.StudentsLesson"
                            }
                        }
                    }
                }
            }
        },
        "/{education_form}/{department}/{group_number}/subgroups": {
            "get": {
                "description": "Retrieves the subgroups list of a group based on department, education form and group number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "groups"
                ],
                "summary": "get a list of subgroups of a certain group",
                "operationId": "get-group-subgroups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Education form, e.g. `do`",
                        "name": "education_form",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Department URL, e.g. `knt` for CSIT department",
                        "name": "department",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Group number, e.g. `351`",
                        "name": "group_number",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Subgroups"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Departments": {
            "type": "object",
            "properties": {
                "full_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "short_name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "model.Groups": {
            "type": "object",
            "properties": {
                "department_id": {
                    "type": "integer"
                },
                "education_form": {
                    "type": "string"
                },
                "group_num": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                }
            }
        },
        "model.StudentsLesson": {
            "type": "object",
            "properties": {
                "day_num": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                },
                "lesson_place": {
                    "type": "string"
                },
                "lesson_type": {
                    "type": "string"
                },
                "subgroup_name": {
                    "type": "string"
                },
                "teacher": {
                    "type": "string"
                },
                "week_type": {
                    "type": "string"
                }
            }
        },
        "model.Subgroups": {
            "type": "object",
            "properties": {
                "group_id": {
                    "type": "integer"
                },
                "subgroup_name": {
                    "type": "string"
                }
            }
        }
    }
}